[refactor] 컴포저블 함수 추출
