+++ core/ui/src/main/java/com/linkedlist/linkllet/core/ui/LnkSearchBar.kt
@@ -1,0 +1,63 @@
+package com.linkedlist.linkllet.core.ui
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.linkedlist.linkllet.core.designsystem.icon.LnkIcon
+import com.linkedlist.linkllet.core.designsystem.icon.lnkicon.Search
+import com.linkedlist.linkllet.core.designsystem.theme.Gray100
+
+@Composable
+fun SearchBar(
+    modifier: Modifier = Modifier,
+    value: String,
+    onValueChange: (String) -> Unit,
+    search: () -> Unit,
+) {
+    Box(
+        modifier = modifier.height(50.dp),
+        contentAlignment = Alignment.CenterEnd,
+    ) {
+        Surface(
+            Modifier.fillMaxHeight(),
+            shape = RoundedCornerShape(12.dp),
+            color = Gray100,
+        ) {
+            BasicTextField(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(start = 20.dp, end = 54.dp),
+                value = value,
+                onValueChange = onValueChange,
+                singleLine = true,
+                textStyle = TextStyle(fontSize = 14.sp, color = Color.Black),
+                decorationBox = { innerTextField ->
+                    Box(contentAlignment = Alignment.CenterStart) {
+                        innerTextField()
+                    }
+                }
+            )
+        }
+        Icon(
+            modifier = Modifier
+                .padding(end = 16.dp)
+                .clickable { search() },
+            imageVector = LnkIcon.Search,
+            contentDescription = "검색",
+        )
+    }
+}
